% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redesim_data_uf.R
\name{redesim_fetch}
\alias{redesim_fetch}
\title{Carrega dados da redesim}
\usage{
redesim_fetch(ano = 2021, mes = 3, uf = "")
}
\arguments{
\item{ano}{ano de consulta. Por padrão, 2021.}

\item{mes}{mes de consulta. Por padrão, março (mês 3).}

\item{uf}{UF de consulta. Por padrão, baixa de todo o Brasil.}
}
\description{
Carrega dados da redesim
}
